Support .mdastrc structure
The oringinal implementation was assuming a user's .mdastrc only had the
config for the mdast-lint plugin. This worked, but it means you could
not use the same file with the `mdast` CLI. This change adds support for
the full structure, so the same file can be used with the CLI tool and
on Hound.

That structure is:

   {
     "plugins": {
       "lint": {
         "rule1": options,
         "rule2": options,
       }
     }
   }

https://github.com/wooorm/mdast-lint#configuring-mdast-lint